Move users db to internal
